+++ Tests/FluentAssertions.Specs/Exceptions/ThrowAssertionsSpecs.cs
@@ -7,6 +7,16 @@ namespace FluentAssertions.Specs.Exceptions;
 public class ThrowAssertionsSpecs
 {
     [Fact]
+    public void When_subject_throws_any_exception_it_should_not_do_anything()
+    {
+        // Arrange
+        Does testSubject = Does.Throw<InvalidOperationException>();
+
+        // Act / Assert
+        testSubject.Invoking(x => x.Do()).Should().Throw();
+    }
+
+    [Fact]
     public void When_subject_throws_expected_exception_it_should_not_do_anything()
     {
         // Arrange
@@ -17,6 +27,16 @@ public class ThrowAssertionsSpecs
     }
 
     [Fact]
+    public void When_func_throws_any_exception_it_should_not_do_anything()
+    {
+        // Arrange
+        Does testSubject = Does.Throw<InvalidOperationException>();
+
+        // Act / Assert
+        testSubject.Invoking(x => x.Return()).Should().Throw();
+    }
+
+    [Fact]
     public void When_func_throws_expected_exception_it_should_not_do_anything()
     {
         // Arrange
@@ -24,6 +44,16 @@ public class ThrowAssertionsSpecs
 
         // Act / Assert
         testSubject.Invoking(x => x.Return()).Should().Throw<InvalidOperationException>();
+    }
+
+    [Fact]
+    public void When_action_throws_any_exception_it_should_not_do_anything()
+    {
+        // Arrange
+        var act = new Action(() => throw new InvalidOperationException("Some exception"));
+
+        // Act / Assert
+        act.Should().Throw();
     }
 
     [Fact]